Initial commit of existing code.
Existing code features some C++ source code and GLSL shaders for
tests and display.  Also a makefile and sample input is provided.
Note that this is the beginning of a major rework since the
initial tests worked but were not flexible enough to support a
full application.  The Scene, Item, and Box classes, as well as
Application and Design are pretty good, but the layers in between
are non-existent or scattered in old code.  Will be working on
Command and Interpreter next.