Mouse now uses Binding.  Started work on visual Manipulators.
Maybe need to add capability to Binding to take a pointer to a
Vector as an argument.  That way it would be very easy to handle
GRAB and Manipulators.  It also might be valuable to have Mouse
always check what's under the cursor when it clicks.  This may
finally make the picking code move out to a Picker class, which
is probably a good thing anyway.  Finally, support for the
movement in the translator manipulator is only in testing.  Need
to figure out how to invert the Projection and Modelview matrices
so they can be used to alter the mouse difference vector, which
would then be applied to the objects.