Created Traits module for interfaces/base classes.
Also moved common settings and actions into 'module.mak' in the
app directory, which simplifies the module makefiles
considerably.  Lastly, created a bunch of test programs that are
easily compiled with 'make tests' or 'make CLASS-test', where
CLASS is the name of class that has a test made for it.