Creation in Factory. XML file supports Uniform.
Also made Tag more useful by adding 'get' methods that look up
an attribute by name and set a reference variable to the value
of the attribute.  If the attribute does not exist in the tag,
an exception is thrown listing the name of the tag and
attribute.  Currently the exception is caught by Producer,
which will print the exception and exit.