Painter draws outline/manipulators correctly.
Now that we're using blending, it's important to draw items in
the correct order.  Besides just drawing the shapes themselves in
order, we also have to worry about any UI elements we add, such
as the outline and translators that we've been using.

Originally we were drawing the UI elements with the shapes, so
they were in the correct order, but then we realized that the
elements were getting the shaders applied to them.  To fix that
we moved drawing the elements to after all the nodes, which means
any program nodes would have removed themselves, and the UI
elements were drawn correctly with the fixed-functionality.

However, this presented a new problem that went unnoticed, in
that the UI elements were not being blended correctly since they
were being drawn out of order.

To fix both the issues, I added capability to Program to
temporarily disable shading.  Now the UI elements are drawn
with fixed-functionality and are blended correctly.

Also, I cleaned up the outline by using a Box drawn in wireframe
mode, and the translator by only drawing the part outside of the
outline, as well as adding a disc to the back side of the cone.