Picker now chooses the closest shape again.
At one point Picker did pick the closest shape, but then Item
was changed to Identifiable with all the interfaces, and it just
didn't get changed again.  Now that we're sorting the scene by
depth because of the blending, it made it quite easy to put this
back in.  We're just using "Node::getDepth", although
manipulators are still always chosen before shapes.