Moved Parser into Factory.
Scene just passes its filename and root node to Factory.  Then
Factory uses Parser to get the Tags.  The process method, which
required you to pass the vector of Tags, is now replaced by start,
which takes no arguments.  Not only does this simplify things a
little bit, but it clears the way for the Factory to use file paths
that are relative to the scene file.