Uniform converts samplers to lowercase to compare.
Previously if user typed "sampler3D" or "sampler2D" it wouldn't
work.  Should probably move this functionality to Tag though and
have a parameter that turns it off.  The only reason it isn't
that way is because filenames might be case sensitive.  (Tag
already converts tag names and attribute keys to lowercase.)