Extended get capabilities in Dataset.
Added get() w/o type parameter.  Added getAsByte(), getAsFloat(),
and getAsShort() methods, which find the position in the data
specified by the index and dereference the void pointer properly,
as well as returning a copy of the value.

Generally, to use it, get the type first, then put that in a switch
statement with the appropriate getAs method in the cases.  For
instance, the GL_UNSIGNED_BYTE case would be getAsByte(index).