Lighting shader calculates direction to light.
Before it was assuming the light position was the direction.  Will
have to think about that because for volumetric it would be better
to have it be the direction.  Oh, and just so I remember, you
make it be a direction by passing it a position with a w component
of 0, rather than 1.  Could maybe have a direction node that takes
care of that, and the user would just have to make sure he or she
does the right thing in the shader.