Preprocessor adds newlines!  Arg!
GLSL does not necessarily treat each source string as a line, so
it doesn't add newlines between them.  This was messing up the
version pragma, as well as the others I tried before.  Although
include still doesn't work so it is still kind of hard to get away
from not having some type of preprocessor, and to do include it is
still good to have defines implemented, so I guess it wasn't a
waste of time, but arg!