Moved opening back to Scene.
Motivation was that Scene is the one that knows about the file,
whereas Factory should just make nodes without much regard to an
actual file.  (Although we do optionally pass it to Factory to
filter "file" attributes.)

This also allowed us to simplify setRoot a little since it can now
be called before the scene is opened.