Boolean "working" as "AND" operator.
Need to handle texture coordinates correctly, which is going to
be a little tricky.  Also should probably make Boolean a base
class and have BooleanAND and BooleanXOR derive from it to make it
easier to manage.  Lastly probably don't need the list of shapes
in BooleanAND at least, but should still at least count them and
for now check that they're both Cubes.